Separated __init__.py to modules
